+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-### 1.6.9 (20480, 2022-02-21)
+### 1.6.10 (20503, 2022-03-07)
+- Added `_ba.get_client_public_device_uuid` function which returns a semi-permanent device id for a connected client running 1.6.10 or newer. Can be useful to combat spam attacks or other mischief.
+- Fixed an issue with `make update` not properly rewriting Visual Studio project files to account for new/deleted source files.
+- Removed various bits of code associated with the (no-longer-functional) Google Play Games multiplayer connections.
+
+### 1.6.9 (20486, 2022-02-22)
 - Upgraded Android Python to 3.9.10
 - Fixed an issue with SSL in Android builds that was preventing communication with the master-server in 1.6.8
 - Added a new network-diagnostics tool at 'Settings->Advanced->Network Testing'. Can be used to diagnose issues talking to master-servers/etc. (especially useful now that SSL can factor in)